+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed relay node not applying custom conversions when needed
 - Fixed the node create menu not listing all nodes when created from edge drag.
 - Fixed Int vector drawer display in nodes.
+- Fixed parameter not editable in Unity 2021.2.
 
 ### Changed
 - Deprecated the FieldFactory and [FieldDrawer] system. You can use Unity's CustomPropertyDrawer instead.