+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.0]
+
+### Fixed
+- Fixed fields with [Settings] attribute not showing up with inheritance.
+
 ## [1.3.0]
 
 ### Added